send_id is not always included
I reached out to support on this as well.

I'm having some issues with the output postback stuff which we are trying to use to capture optouts from sailthru and sync them back to our database.

According to the [Sailthru docs](https://getstarted.sailthru.com/developers/api-basics/postbacks/)
> If the message is used as double opt-in and the verification is successful, the send_id and email address will be posted to this url.

We aren't using double opt in and we're not getting `send_id` on the postback. So... if its not there don't validate it.